Migrates Elasticsearch test images to OSS license
This uses OSS license for Elasticsearch 6 and 7, reducing the size of
the images. Image size is further reduced in 7 by using a recent change
to our JRE that allows Thai segmentation support (module jdk.localedata)

This also reduces test execution time by only executing the 7.8+ schema
test vs the entire suite again.
