+++ docker/storage/elasticsearch6/config/elasticsearch.yml
@@ -1,5 +1,5 @@
 #
-# Copyright 2015-2019 The OpenZipkin Authors
+# Copyright 2015-2020 The OpenZipkin Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 # in compliance with the License. You may obtain a copy of the License at
@@ -14,4 +14,4 @@
 
 network.host: 0.0.0.0
 discovery.zen.minimum_master_nodes: 1
-xpack.ml.enabled: false
+
