+++ zipkin-storage/elasticsearch/src/test/java/zipkin2/elasticsearch/integration/ITElasticsearchStorageV7.java
@@ -13,10 +13,15 @@
  */
 package zipkin2.elasticsearch.integration;
 
+import java.io.IOException;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.RegisterExtension;
+import zipkin2.elasticsearch.ElasticsearchStorage;
 
-/** For testing legacy template */
+import static zipkin2.elasticsearch.integration.ElasticsearchStorageExtension.index;
+
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class ITElasticsearchStorageV7 extends ITElasticsearchStorage {
 
@@ -25,5 +30,16 @@ class ITElasticsearchStorageV7 extends ITElasticsearchStorage {
 
   @Override ElasticsearchStorageExtension backend() {
     return backend;
+  }
+
+  @Nested
+  class ITEnsureIndexTemplate extends zipkin2.elasticsearch.integration.ITEnsureIndexTemplate {
+    @Override protected ElasticsearchStorage.Builder newStorageBuilder(TestInfo testInfo) {
+      return backend().computeStorageBuilder().index(index(testInfo));
+    }
+
+    @Override public void clear() throws IOException {
+      storage.clear();
+    }
   }
 }
